Move SdkConfig out of MeticaAPI
and make MeticaAPI not a ScriptableObject
