+++ SDK/Runtime/SdkConfig.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d428ff52d901f324fb0d70b066619492
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
